requirements.txt
django==3.1.13
mssql-django==1.1.3
pyodbc==4.0.34
djangorestframework==3.11.2
django-cors-headers==3.13.0